requirements.txt
certifi
charset-normalizer
idna
lxml
PyExecJS
PyMuPDF
python-docx
requests
six
tqdm
urllib3
openai
tenacity
